First step in refactoring.
